Add files via upload
This project is build using HTML, CSS, Bootstrap and JavaScript. This project stores the data without using the database i.e, it uses the local storage to absorb the data given by the user. The use of JavaScript makes this project very attractive and dynamic such as it throws the message whenever a user tries to submit an empty field and throws the message whenever the user adds or deletes the book.